+++ freecad/gridfinity_workbench/feature_construction.py
@@ -954,92 +954,6 @@ def make_bin_base(obj: FreeCAD.DocumentObject, layout) -> Part.Shape:
 
     return assembly2
 
-
-def make_baseplate_center_cut(obj: FreeCAD.DocumentObject) -> Part.Shape:
-    """Create baseplate center cutout.
-
-    Args:
-        obj (FreeCAD.DocumentObject): Document object.
-
-    Returns:
-        Part.Shape: Baseplate center cutout shape.
-
-    """
-    inframedis = (
-        obj.GridSize / 2
-        - obj.BaseProfileTopChamfer
-        - obj.BaseProfileBottomChamfer
-        - obj.BaseplateTopLedgeWidth
-    )
-    magedge = (
-        obj.GridSize / 2
-        - obj.MagnetHoleDistanceFromEdge
-        - obj.MagnetHoleDiameter / 2
-        - obj.MagnetEdgeThickness
-    )
-    magcenter = obj.GridSize / 2 - obj.MagnetHoleDistanceFromEdge
-    smfillpos = inframedis - obj.SmallFillet + obj.SmallFillet * math.sin(math.pi / 4)
-    smfillposmag = magedge - obj.SmallFillet + obj.SmallFillet * math.sin(math.pi / 4)
-    smfilloffcen = (
-        obj.GridSize / 2
-        - obj.MagnetHoleDistanceFromEdge
-        - obj.MagnetHoleDiameter / 2
-        - obj.MagnetEdgeThickness
-        - obj.SmallFillet
-    )
-    smfillins = inframedis - obj.SmallFillet
-    bigfillpos = (
-        obj.GridSize / 2
-        - obj.MagnetHoleDistanceFromEdge
-        - (obj.MagnetHoleDiameter / 2 + obj.MagnetEdgeThickness) * math.sin(math.pi / 4)
-    )
-    mec_middle = FreeCAD.Vector(0, 0, 0)
-
-    v1 = FreeCAD.Vector(0, -inframedis, 0)
-    v2 = FreeCAD.Vector(-smfilloffcen, -inframedis, 0)
-    v3 = FreeCAD.Vector(-magedge, -smfillins, 0)
-    v4 = FreeCAD.Vector(-magedge, -magcenter, 0)
-    v5 = FreeCAD.Vector(-magcenter, -magedge, 0)
-    v6 = FreeCAD.Vector(-smfillins, -magedge, 0)
-    v7 = FreeCAD.Vector(-inframedis, -smfilloffcen, 0)
-    v8 = FreeCAD.Vector(-inframedis, 0, 0)
-
-    va1 = FreeCAD.Vector(-smfillposmag, -smfillpos, 0)
-    va2 = FreeCAD.Vector(-bigfillpos, -bigfillpos, 0)
-    va3 = FreeCAD.Vector(-smfillpos, -smfillposmag, 0)
-
-    l1 = Part.LineSegment(v1, v2)
-    ar1 = Part.Arc(l1.EndPoint, va1, v3)
-    l2 = Part.LineSegment(ar1.EndPoint, v4)
-    ar2 = Part.Arc(l2.EndPoint, va2, v5)
-    l3 = Part.LineSegment(ar2.EndPoint, v6)
-    ar3 = Part.Arc(l3.EndPoint, va3, v7)
-    l4 = Part.LineSegment(ar3.EndPoint, v8)
-    l5 = Part.LineSegment(l4.EndPoint, mec_middle)
-    l6 = Part.LineSegment(l5.EndPoint, l1.StartPoint)
-
-    wire = Utils.curve_to_wire([l1, ar1, l2, ar2, l3, ar3, l4, l5, l6])
-    partial_shape1 = Part.Face(wire).extrude(FreeCAD.Vector(0, 0, -obj.TotalHeight))
-    partial_shape2 = partial_shape1.mirror(FreeCAD.Vector(0, 0, 0), FreeCAD.Vector(0, 1, 0))
-    partial_shape3 = partial_shape1.mirror(FreeCAD.Vector(0, 0, 0), FreeCAD.Vector(1, 0, 0))
-    partial_shape4 = partial_shape2.mirror(FreeCAD.Vector(0, 0, 0), FreeCAD.Vector(1, 0, 0))
-
-    shape = partial_shape1.multiFuse([partial_shape2, partial_shape3, partial_shape4])
-
-    vec_list: list[FreeCAD.Vector] = []
-    xtranslate = 0
-    ytranslate = 0
-
-    for _ in range(obj.xGridUnits):
-        ytranslate = 0
-        for _ in range(obj.yGridUnits):
-            vec_list.append(FreeCAD.Vector(xtranslate, ytranslate, 0))
-            ytranslate += obj.GridSize.Value
-        xtranslate += obj.GridSize.Value
-
-    return Utils.copy_and_translate(shape, vec_list)
-
-
 def make_bottom_hole_shape(obj: FreeCAD.DocumentObject) -> Part.Shape:
     """Create bottom hole shape.
